examples: Fix memory layout in MPS3_CM55_DTCM.sct
Fix the memory layout defined in MPS3_CM55_DTCM.sct which is used for
Corstone-300:
* The sizes of ROM (ITCM) and RAM (DTCM) should be 0x80000 bytes each,
much less than previously declared. Any code beyond the intended ROM
size can't be loaded onto the target but nonetheless got linked into the
image previously due to incorrect size declared in the linker script.
* As a result of the ROM size fix, the ROM is not enough to fit the
example anymore. To fix this, add an additional region ROM region which
is the code SRAM and put read-only code and data in it. The use of the
two ROM regions, ITCM and code SRAM, is consistent with other projects
such as TF-M.

Signed-off-by: Lingkai Dong <lingkai.dong@arm.com>